+++ docs/integration-test-helpers/masonry/ExampleGridItem.tsx
@@ -23,15 +23,12 @@ export default function ExampleGridItem({ data = {}, itemIdx, expanded }: Props)
   const isTwoColItem = data.columnSpan === 2;
 
   return (
-    <div
-      style={{
-        padding: '0 7px 14px',
-      }}
-    >
+    <div>
       <div
+        className="grid-item-test"
         style={{
           height: expanded ? data.height + 100 : data.height,
-          border: '1px solid #ff0000',
+          boxSizing: 'border-box',
           background: isTwoColItem ? 'black' : data.color,
           color: isTwoColItem ? 'white' : undefined,
         }}
